Fixed a bug that caused the desktop sidebar to close when hiding a manager component.
Prior to this fix, hiding a component in the manager on the desktop version of the application would also close the sidebar, making it difficult for the user to navigate. After this fix, the sidebar will no longer close when hiding a component in the manager. The JavaScript code has been updated to implement this fix. With this correction, users of the desktop version of the application will be able to navigate more effectively and efficiently.